requirements.txt
aioredis==2.0.0
Quart==0.15.1
ffmpeg-python==0.2.0
redis==3.5.3
Quart-CORS==0.5.0
